rag_pdf/requirements.txt
langchain-huggingface
langchain-core
langchain
langchain-community
faiss-cpu
python-dotenv
transformers
sentence-transformers
unstructured
pdf2image
pdfminer.six
pi-heif
